requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.7
django-oauth-toolkit==2.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==1.3.0
ecdsa==0.18.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
urllib3==1.26.15
wrapt==1.15.0
